Allow previous nodes to be re-selected
Ditched redux-undo in favor of a home-grown solution; the Graph
duck has a `history` prop that holds a recursive set of states:

history: {
  vertices: [4],
  edges: [4],
  history: {
    vertices: [3],
    edges: [3],
    history: {
      vertices: [2],
      edges: [2],
      history: {
        vertices: [1],
        edges: [1]
      }
    }
  }
}

The advantage to this solution is simplicity. I can literally just
call `state.get('history')` to restore the previous state, as many
times as needed. Because of Immutable.js' structural sharing, I
don't believe it's even that wasteful :)