Create artist placeholder, expand action
Dispatching a selectArtist action now only selectively adds the item
to the history if a boolean flag is set to true (which is the default).
That way, on the initial load with a custom artistId query param, we
aren't needlessly setting it to the value we just read.

Also created an artist placeholder, so that the action is always
receiving an immutable Map instead of an artistId occasionally. Yay
type safety!