Refactor updateArtistUrl to be more generalized
Instead of making the 'Start Over' button trigger all the restart
stuff directly, I thought it'd be better to make it just update the
URL and let the history enhancer handle all the actions and stuff.

This required refactoring `updateArtistUrl` into a more generalized
`updateUrl` method, which defaults to the homepage when no artistId
is supplied.

This is better, because it means all related actions are passing
through the same system; the "Start Over" button is really just a
link, and it works exactly how the vertices and the back/forwards
buttons do.